Move RDHAny from DataFormats/Headers to Detectors/Raw
Asked by Gvozden to avoid DataFormats/Headers dependence on ROOT and GPU headers
